chore: move all endpoint files into a simplier 'routers' files + update Models endpoint to adapt the new fix
